Update precision to int4 in pipeline

+++ src/run_pipeline.py
@@ -33,7 +33,7 @@ except Exception:
 
 MODELS = [
     {"name": "Gemma-3-4B", "id": "google/gemma-3-4b-it", "precision": "fp16"},
-    {"name": "Gemma-3-12B", "id": "google/gemma-3-12b-it", "precision": "int8"},
+    {"name": "Gemma-3-12B", "id": "google/gemma-3-12b-it", "precision": "int4"},
 ]
